Upgrade lessons 1-8 to vulkano version 0.31.0
In addition to keeping up with the ever-changing vulkano API, this allows these lessons to run out of the box on macOS, using MoltenVK,
by explicitly setting the enumerate_portability flag.  This flag does nothing on other platforms.
(In fact, the vulkano team is considering making this flag true by default in future versions of vulkano.)

A few small refactors and textual edits for readability.

Consistently use UNIX-style line breaks, including in .toml files.

The intent is to finish porting all of the lessons soon...