Fix issue with radians and perspective()
It was really bugging me that some of the perspective looked wrong.
I finally found the root cause.  The glm::perspective() function takes radians, not degrees.
Use glm::half_pi() as the fovy, which is a relatively wide angle of 90 degrees as the 'fovy'.

Also, the 'up' vector for the view was being passed as vec(0.0, -1.0, 0.0),
which compensated for the wraparound distortion caused by using 180.0 for the fovy.

The tutorial talks about positive Y as pointing down in Vulkan, and all of the models assume this.

Fixing the call to perspective and changing the up vector back to 1.0 returns things to normal.
Had to move a few objects back in the negative Z direction, to move away from the camera.

Added an initial rotation by pi around the Y-axis to the model in lesson_9, so that the suzanne monkey starts out by facing the camera.