Add lesson_13
According to the vulkano-0.32 changelog:

```
DeviceLocalBuffer::{from_buffer, from_data, from_iter} and ImmutableImage::{from_iter, from_buffer} now take a mutable reference to an AutoCommandBufferBuilder instead of a queue, and no longer return a future. The upload command will be recorded into the provided command buffer, which should be executed later.
```

Therefore, we still load the image from the file before the event loop, but we don't load the image into the GPU texture until we execute the first frame.
We could also move the file loading into the event loop, in order to avoid keeping the image data in CPU memory, but that's unnecessary for this example.