Vulkano 0.32.0 (#4)
* Upgrade lessons 1-8 to vulkano version 0.31.0

In addition to keeping up with the ever-changing vulkano API, this allows these lessons to run out of the box on macOS, using MoltenVK,
by explicitly setting the enumerate_portability flag.  This flag does nothing on other platforms.
(In fact, the vulkano team is considering making this flag true by default in future versions of vulkano.)

A few small refactors and textual edits for readability.

Consistently use UNIX-style line breaks, including in .toml files.

The intent is to finish porting all of the lessons soon...

* Convert to vulkano-0.32.0

* Add lessons 9-10

* Add lesson_11

* Move buffer creation to after pipeline in every lesson, for consistency

* Fix issue with radians and perspective()

It was really bugging me that some of the perspective looked wrong.
I finally found the root cause.  The glm::perspective() function takes radians, not degrees.
Use glm::half_pi() as the fovy, which is a relatively wide angle of 90 degrees as the 'fovy'.

Also, the 'up' vector for the view was being passed as vec(0.0, -1.0, 0.0),
which compensated for the wraparound distortion caused by using 180.0 for the fovy.

The tutorial talks about positive Y as pointing down in Vulkan, and all of the models assume this.

Fixing the call to perspective and changing the up vector back to 1.0 returns things to normal.
Had to move a few objects back in the negative Z direction, to move away from the camera.

Added an initial rotation by pi around the Y-axis to the model in lesson_9, so that the suzanne monkey starts out by facing the camera.

* Use much simpler Option::as_mut() to mutably append commands

Should probably update the tutorial text to mention this technique.

* Add lesson_11_5

* Add lesson_12

* Shorten all _subbuffer variable names, both for consistency/readability, and to avoid line wrapping in cargo fmt

* Fix input_attachment_index to match the setup of the render_pass

* Upgrade to vulkano-0.32.1 patch.  No API changes necessary

* Add lesson_13

According to the vulkano-0.32 changelog:

```
DeviceLocalBuffer::{from_buffer, from_data, from_iter} and ImmutableImage::{from_iter, from_buffer} now take a mutable reference to an AutoCommandBufferBuilder instead of a queue, and no longer return a future. The upload command will be recorded into the provided command buffer, which should be executed later.
```

Therefore, we still load the image from the file before the event loop, but we don't load the image into the GPU texture until we execute the first frame.
We could also move the file loading into the event loop, in order to avoid keeping the image data in CPU memory, but that's unnecessary for this example.

* Add lesson_14