Update docs for lesson_3
I removed much of the discussion about the triangle being upside-down, and updated the corresponding image.
The upside-down triangle was an artifact of the previous fix to the `fovy` being passed as degrees to `glm::perspective()` instead of radians.
By using radians, you won't ever get an upside-down triangle, and we don't need to then flip the y-direction again as part of the view matrix.