Update docs for lesson_10
Reorder the section about refactoring `vp_set` and `vp_layout`.  There must have been some previous changes (before the vulkano-0.32.0 changes), which introduced some inconsistencies in that section.  Just reordering and removing the redundant code seems to make it make sense again, and reads much easier.

Add the same rotate by pi that we introduced in lesson_9.  It doesn't actually change anything, because the model is symmetric when flipped, but keep it for consistency with the previous lesson.