Add docs for lesson_13
vulkano-0.32.0 changed the way that textures are asynchronously loaded.
An ImmutableImage::from_iter() call now takes a reference to a command builder, instead of returning a GpuFuture.

Therefore, the code itself was refactored to laod on the first frame.  (That was done in the vulkano-0.32.0 PR).
This is now describing that in more detail in the lesson_13 docs.