Use a `Cell<Option<T>>` to cache the model matrices, instead of a boolean `requires_update` flag
In addition to being clear about which fields are actually cached, it also allows the
`model_matrices` method to use `&self` instead of `&mut self` in the signature,
and makes it clear that the caching is an internal concern of the Model.

This bubbles all the way to the public API for the System.  See the change to `System::geometry()`, for example.

It's also a good introduction to the concept of interior mutability in Rust.
We may want to add a small section to the docs about this.

Also fix mention of obj_loader in section_9.md